Changed monitor creation to sets
This commit changes the monitor list into a set, allowing for an easier
interface.

It changes populate_framebuffer so as to fit into a ws_set_select call. As this
is the way to iterate over the set.

Find connector now also searches using a dummy object instead of iterating
over a linked list.